Added storage-based flood
Creates an array at `mclg:compile run`` of input / gate / output blocks in reverse topological order (starting bits are at the end of the array).
Each index in the array has specific data tied to it, the most important of which being it's input and output bitpack and bitmasks, as well as it's type.
Updated the `toggle_bit` function to also update the initial bit pack.