Added truth table compiling and combining
Yes, the system is kinda jank. Will make more intuitve and faster later, but it works which is what matters
Use the runtime manager to toggle bits on and off, select an area to compile that area. Run "name_truth_table" to name the truth table.
You need to reassign truth tables to the selected area after creating them.
Run "compile_complete" to combine the truth tables.